requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.0.5
django-constance==2.8.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.1
sqlparse==0.4.2
wrapt==1.14.1
django-sendmail-backend==0.1.2